+++ packages/utils/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Removed
+
+- Removed `getIndentation`, `setDefaultTabWidth`, and `getDefaultTabWidth` helpers
 
 ## [16.0.8] - 2026-06-18